Seperate commands enum from parser
